+++ CHANGELOG.md
@@ -14,9 +14,20 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Features
 
-- plugin: add feature `plugin` ([#129](https://github.com/lamports-dev/richat/pull/129))
-
 ### Breaking
+
+## 2025-07-22
+
+- richat-plugin-agave-v4.0.1
+- richat-v4.0.1
+
+### Fixes
+
+- richat: fix allocation for entries during dedup ([#130](https://github.com/lamports-dev/richat/pull/130))
+
+### Features
+
+- plugin: add feature `plugin` ([#129](https://github.com/lamports-dev/richat/pull/129))
 
 ## 2025-07-10